Allow defaults of differing type in env parsers
