+++ runtests.py
@@ -1,14 +1,15 @@
 #!/usr/bin/env python
 import os
 import sys
+from typing import List, Optional
 
 import django
 from django.conf import settings
 from django.test.utils import get_runner
 
 
-def main(args=None):
-    os.environ['DJANGO_SETTINGS_MODULE'] = 'tests.settings'
+def main(args: Optional[List[str]] = None) -> None:
+    os.environ["DJANGO_SETTINGS_MODULE"] = "tests.settings"
     django.setup()
     TestRunner = get_runner(settings)
     test_runner = TestRunner(verbosity=2)
