+++ setup.cfg
@@ -104,12 +104,6 @@ disallow_incomplete_defs = True
 disallow_untyped_defs = True
 ignore_missing_imports = False
 
-# Initially ignore errors by default, and only fail for the modules marked
-# below. Once we cover a significant portion of the codebase we should switch
-# strategy to raise errors by default and mark modules that we haven't covered
-# yet.
-ignore_errors = True
-
 warn_unused_configs = True
 warn_redundant_casts = True
 warn_unused_ignores = True
@@ -122,8 +116,16 @@ plugins =
 [mypy.plugins.django-stubs]
 django_settings_module = tests.settings
 
+# Initially ignore errors by default, and only fail for the modules marked
+# below. Once we cover a significant portion of the codebase we should switch
+# strategy to raise errors by default and mark modules that we haven't covered
+# yet.
+[mypy-bananas.*]
+ignore_errors = True
+
 [mypy-tests.*]
 disallow_untyped_defs = False
+ignore_errors = True
 
 [mypy-bananas.url]
 ignore_errors = False
