+++ tests/test_env.yaml
@@ -1,0 +1,69 @@
+- case: env_get_bool_overloaded_signatures
+  main: |
+    from bananas.environment import env
+    reveal_type(env.get_bool("test"))
+    reveal_type(env.get_bool("test", False))
+    reveal_type(env.get_bool("test", "false"))
+  out: |
+    main:2: note: Revealed type is 'Union[builtins.bool, None]'
+    main:3: note: Revealed type is 'builtins.bool'
+    main:4: note: Revealed type is 'Union[builtins.bool, builtins.str*]'
+  mypy_config: |
+    [mypy-django.*]
+    ignore_errors = True
+
+- case: env_get_int_overloaded_signatures
+  main: |
+    from bananas.environment import env
+    reveal_type(env.get_int("test"))
+    reveal_type(env.get_int("test", 1))
+    reveal_type(env.get_int("test", "1"))
+  out: |
+    main:2: note: Revealed type is 'Union[builtins.int, None]'
+    main:3: note: Revealed type is 'builtins.int'
+    main:4: note: Revealed type is 'Union[builtins.int, builtins.str*]'
+  mypy_config: |
+    [mypy-django.*]
+    ignore_errors = True
+
+- case: env_get_tuple_overloaded_signatures
+  main: |
+    from bananas.environment import env
+    reveal_type(env.get_tuple("test"))
+    reveal_type(env.get_tuple("test", ()))
+    reveal_type(env.get_tuple("test", []))
+  out: |
+    main:2: note: Revealed type is 'Union[builtins.tuple[builtins.str], None]'
+    main:3: note: Revealed type is 'builtins.tuple[builtins.str]'
+    main:4: note: Revealed type is 'Union[builtins.tuple[builtins.str], builtins.list*[<nothing>]]'
+  mypy_config: |
+    [mypy-django.*]
+    ignore_errors = True
+
+- case: env_get_list_overloaded_signatures
+  main: |
+    from bananas.environment import env
+    reveal_type(env.get_list("test"))
+    reveal_type(env.get_list("test", ["strs"]))
+    reveal_type(env.get_list("test", ()))
+  out: |
+    main:2: note: Revealed type is 'Union[builtins.list[builtins.str], None]'
+    main:3: note: Revealed type is 'builtins.list[builtins.str]'
+    main:4: note: Revealed type is 'Union[builtins.list[builtins.str], Tuple[]]'
+  mypy_config: |
+    [mypy-django.*]
+    ignore_errors = True
+
+- case: env_get_set_overloaded_signatures
+  main: |
+    from bananas.environment import env
+    reveal_type(env.get_set("test"))
+    reveal_type(env.get_set("test", {"doot"}))
+    reveal_type(env.get_set("test", ()))
+  out: |
+    main:2: note: Revealed type is 'Union[builtins.set[builtins.str], None]'
+    main:3: note: Revealed type is 'builtins.set[builtins.str]'
+    main:4: note: Revealed type is 'Union[builtins.set[builtins.str], Tuple[]]'
+  mypy_config: |
+    [mypy-django.*]
+    ignore_errors = True
