+++ tests/tests.py
@@ -244,61 +244,6 @@ class QuerySetTest(TestCase):
         md = ModelDict(on=False, monkey__name="Pato", monkey__banana=True)
         md.expand()
         self.assertEqual(md, {"on": False, "monkey": {"name": "Pato", "banana": True}})
-
-
-class EnvTest(TestCase):
-    def test_parse_bool(self):
-        self.assertTrue(environment.parse_bool("True"))
-        self.assertTrue(environment.parse_bool("true"))
-        self.assertTrue(environment.parse_bool("TRUE"))
-        self.assertTrue(environment.parse_bool("yes"))
-        self.assertTrue(environment.parse_bool("1"))
-        self.assertFalse(environment.parse_bool("False"))
-        self.assertFalse(environment.parse_bool("0"))
-        self.assertRaises(ValueError, environment.parse_bool, "foo")
-
-    def test_parse_int(self):
-        self.assertEqual(environment.parse_int("123"), 123)
-        self.assertEqual(environment.parse_int("0644"), 420)
-        self.assertEqual(environment.parse_int("0o644"), 420)
-        self.assertEqual(environment.parse_int("0"), 0)
-
-    def test_parse_tuple(self):
-        self.assertTupleEqual(environment.parse_tuple("a"), ("a",))
-        self.assertTupleEqual(environment.parse_tuple(" a,b, c "), ("a", "b", "c"))
-
-    def test_parse_list(self):
-        self.assertListEqual(environment.parse_list("a, b, c"), ["a", "b", "c"])
-
-    def test_parse_set(self):
-        self.assertSetEqual(environment.parse_set("b, a, c"), {"a", "b", "c"})
-
-    def test_env_wrapper(self):
-        self.assertEqual(env.get("foo", "bar"), "bar")
-
-        self.assertEqual(env.get("foo", "bar"), "bar")
-
-        self.assertIsNone(env.get_bool("foobar"))
-
-        self.assertFalse(env.get_bool("foobar", False))
-        environ["foobar"] = "True"
-        self.assertTrue(env.get_bool("foobar", False))
-        environ["foobar"] = "Ture"
-        self.assertIsNone(env.get_bool("foobar"))
-        self.assertFalse(env.get_bool("foobar", False))
-
-        environ["foobar"] = "123"
-        self.assertEqual(env.get_int("foobar"), 123)
-
-        environ["foobar"] = "a, b, c"
-        self.assertTupleEqual(env.get_tuple("foobar"), ("a", "b", "c"))
-
-        environ["foobar"] = "a, b, c"
-        self.assertListEqual(env.get_list("foobar"), ["a", "b", "c"])
-
-        environ["foobar"] = "a, b, c"
-        self.assertSetEqual(env.get_set("foobar"), {"a", "b", "c"})
-        self.assertEqual(env.get("foobar"), "a, b, c")
         self.assertEqual(env["foobar"], "a, b, c")
